Add watch demo button to the header
